Cache downloads and database creation in a volume, better logging on rest api

+++ app/01-download-gtfs-file.sh
@@ -1,13 +1,37 @@
-#!/bin/bash -e
+#!/usr/bin/env bash
+set -e
 
 url="$( python get-data-url.py )"
+workdir="/store/$( basename "$url" )"
 
-wget -O raw-data.zip "$url"
+echo
+echo "Creating symlink data -> $workdir/data"
+ln -s "$workdir/data" data
 
-rm -rf data
+if [ ! -f "$workdir/data/.data-prepared" ]; then
+  
+  mkdir -p "$workdir"
+  cd "$workdir"
 
-unzip raw-data.zip -d data
+  echo
+  echo "Downloading $url to $workdir/raw-data.zip" 
+  wget -O raw-data.zip "$url"
 
-rm raw-data.zip
+  echo
+  echo "Extracting $workdir/raw-data.zip to $workdir/data" 
+  rm -rf data
+  unzip raw-data.zip -d data
+  rm raw-data.zip
+  
+  echo
+  echo "Performing dos2unix on $workdir/data/*.txt" 
+  dos2unix data/*.txt
+  
+  touch data/.data-prepared
+else
+  
+  echo
+  echo 'Data already downloaded and extracted'
+fi
 
-dos2unix data/*.txt
+
